% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_spatial.R
\name{create_strips}
\alias{create_strips}
\title{create strips}
\usage{
create_strips(field, plot_heading, plot_width, radius)
}
\arguments{
\item{field}{er}

\item{plot_heading}{rtge4}

\item{plot_width}{tgrt4}

\item{radius}{getrg}
}
\description{
temp
}
