% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_buffer}
\alias{make_buffer}
\title{test st_buffer()}
\usage{
make_buffer(sf, radius)
}
\arguments{
\item{sf}{sf}

\item{radius}{radius}
}
\value{
sf
}
\description{
test
}
\examples{

data(nc)
buffer <- make_buffer(sf::st_centroid(nc), 5)
}
