% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_grid}
\alias{make_grid}
\title{test st_make_grid()}
\usage{
make_grid(sf, x_length, y_length)
}
\arguments{
\item{sf}{sf}

\item{x_length}{long length}

\item{y_length}{lat length}
}
\value{
sf
}
\description{
test
}
\examples{

data(nc)
buf <- make_buffer(sf::st_centroid(nc), 100)
grids <- make_grid(buf[1, ], 10, 20)

}
