% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_spatial.R
\name{prepare_ablines}
\alias{prepare_ablines}
\title{create strips}
\usage{
prepare_ablines(ab_line, field, plot_width)
}
\arguments{
\item{ab_line}{er}

\item{field}{rtge4}

\item{plot_width}{tgrt4}
}
\description{
temp
}
