% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{td_single_input}
\alias{td_single_input}
\title{td single input form ofpetrial}
\format{
sf
A data frame with 1 rows and 10 columns:
\describe{
  \item{AREA}{area}
}
}
\usage{
td_single_input
}
\description{
nc as sf
}
\keyword{datasets}
