% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.with.atlas-package.R
\docType{package}
\name{test.with.atlas-package}
\alias{test.with.atlas}
\alias{test.with.atlas-package}
\title{test.with.atlas: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
